Create waveGesture
Uses the light sensor to estimate whether a wave has occurred.

Due to the limitations of the hardware, this won't work if there isn't
enough contrast between the light level when a hand is and isn't over
the sensor.

However, this should still work even when dim, as detection is done via
the rate of change of the light level.